Refactor internal parsing from static analysis
A tremendous amount of effort was having to be maintained to facilitate
static parsing of a codebase for annotations without resorting to
reflection. The effort spent on this didn't have any real value attached
to it outside of choosing when to resolve environment and constant
values. This PR eschews a detailed static analysis of the codebase and
instead uses reflection to determine the appropriate data to gather from
annotations.

This PR also removes the InjectScalar and InjectService annotations as
fixing them to work with this new system would be a fair amount of
effort for something that's being removed with the work on the Inject
annotation.