Improve active profiles API (#92)
Adds an ActiveProfilesBuilder with a fluent API to create a list of
active profiles following the conventions laid out by Annotated
Container. It allows for conditionally adding profiles in addition to
the 'default' profile. The expected use case for this construct is when
you're programatically creating your list of active profiles.

Also adds an ActiveProfilesParser and implementation that parses a CSV
list of active profiles. For example, the string 'default,dev,local'
would result in an array of ['default', 'dev', 'local']. The intended
use case for this construct is when you have the profiles set as a
string in an environment variable or some other hard-coded
configuration.