Remove ability to not share a service
Through more thorough testing of creating non-shared services and
services that depended on other non-shared services it was decied there
was too much confusion and odd behavior around the functionality to
include it in Annotated Container. Without a solid use-case for why
you'd need to do this the idea of a non-shared service should be handled
through the creation of a factory that's marked as a service that is
then depended on to appropriately create new instances.