Add third-party services first
Previously we were running any ContainerDefinitionBuilderContextConsumer
was being ran _after_ we added all the definitions from parsed
Attributes. However, that could result in a situation where a service is
provided by the consumer but then a service delegate or service
preparet that is annotated would fail because the corresponding service
hadn't been added yet. This change moves the addition of third-party
services to happen _first_, before the annotated definitions, ensuring
that any service delegates/prepares properly recognize that there is a
service.

Fixes #192